% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combined_arrays}
\alias{combined_arrays}
\title{Create combined arrays based on selected clusters}
\usage{
combined_arrays(selected_clusters, celltype1, celltype2, celltype3, celltype4)
}
\arguments{
\item{selected_clusters}{A character vector containing the names of selected clusters.}

\item{celltype1}{A matrix containing data for celltype1.}

\item{celltype2}{A matrix containing data for celltype2.}

\item{celltype3}{A matrix containing data for celltype3.}

\item{celltype4}{A matrix containing data for celltype4.}
}
\value{
A list of combined arrays.
}
\description{
This function creates a list of combined arrays based on the selected clusters and corresponding data matrices.
}
