% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_plots_and_save}
\alias{generate_plots_and_save}
\title{Generate plots and save}
\usage{
generate_plots_and_save(
  combined_arrays,
  include_celltypes,
  new_data,
  sim_spatial_spot_true_prop_data,
  prefix = ""
)
}
\arguments{
\item{combined_arrays}{A list containing matrices of data for each cell type}

\item{include_celltypes}{A character vector indicating which cell types to include}
}
\value{
None
}
\description{
This function generates plots of cell locations with a grid and a pie chart of cell type proportions.
The plots are saved as PDF files.
}
\examples{
\dontrun{
generate_plots_and_save(combined_arrays, selected_clusters,
new_data, sim_spatial_spot_true_prop_data, type)
}

}
