% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_type3_data.R
\name{generate_type3_data}
\alias{generate_type3_data}
\title{generate_type3_data}
\usage{
generate_type3_data(meta_data_file, gene_count_file, selected_clusters, Prefix)
}
\arguments{
\item{meta_data_file}{Path to the metadata file.}

\item{gene_count_file}{Path to the gene count file.}

\item{selected_clusters}{A named vector specifying the clusters to include.}

\item{Prefix}{Prefix for the output CSV files.}
}
\description{
This function generates data and plots for selected clusters.
}
\examples{
\dontrun{
Prefix = "type_three"
meta_data_file <- system.file(package="SSTD", "extdata", "meta_data.csv")
gene_count_file <- system.file(package="SSTD", "extdata", "gene_count.csv")
selected_clusters <- c("eL4", "eL5", "eL6", "Endo")
generate_type3_data(meta_data_file, gene_count_file, selected_clusters, Prefix)
}

}
