% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generating_simulation_data.R
\name{Generating_simulation_data}
\alias{Generating_simulation_data}
\title{Generating_simulation_data}
\usage{
Generating_simulation_data(
  spot_spatial_count,
  sim_spatial_spot_loc,
  cell_celltype,
  grid_interval,
  prefix
)
}
\arguments{
\item{spot_spatial_count}{A vector representing the UMI count of each cell.

spot_spatial_count[1:5, 1:5]
  A730017C20Rik Aamp Abat Abhd3 Acat2
0             0    0    1     0     0
1             0    0    1     0     2
2             0    0    0     0     0
3             0    0    0     0     1
4             0    0    0     0     1}

\item{sim_spatial_spot_loc}{A table representing the simulated spatial location.
Each row of the table represents the coordinates of a spatial point, and each column
represents a dimension (such as x, y coordinates).

You must ensure that the column names "X" and "Y" exist in this file.
sim_spatial_spot_loc[1:5, ]
         X         Y
1 24.08588  8867.800
2 26.38492 10853.262
3 49.77125  4980.778
4 53.47311  8729.177
5 38.32856  8978.149}

\item{cell_celltype}{A table representing the type of each cell.Each element represents
the type of a cell, which can be a character vector or factor.

You must ensure that the column names "Cluster" and "celltype" exist in this file.
meta_data[1:5, ]
  Cluster         celltype
0       0     ExcitatoryL4
1       1 ExcitatoryL2and3
2       2     ExcitatoryL6
3       0     ExcitatoryL4
4       0     ExcitatoryL4}

\item{grid_interval}{Spacing size of spatial grids.Used to map the simulated spatial
positions onto the grid for subsequent spatial analysis.}
}
\description{
A function that generates simulated data. This function is used to generate simulated experimental data and
save the results to a file.And the directory path for the output result is the current folder
}
\examples{
\dontrun{
f1 <- system.file(package="SSTD", "extdata", "Spatial_count.txt")
f2 <- system.file(package="SSTD", "extdata", "Locations.txt")
f3 <- system.file(package="SSTD", "extdata", "Spatial_annotate.txt")
Generating_simulation_data(spot_spatial_count = f1,
           sim_spatial_spot_loc = f2,
           cell_celltype = f3,
           grid_interval = 500,
           prefix = "test")
}

}
