% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_gradation_2d}
\alias{get_gradation_2d}
\title{Generate a combined 2D gradation array matrix}
\usage{
get_gradation_2d(start, stop, width, height, is_horizontal)
}
\arguments{
\item{start}{The starting value for the sequence}

\item{stop}{The ending value for the sequence}

\item{width}{The number of columns in the matrix}

\item{height}{The number of rows in the matrix}

\item{is_horizontal}{Logical value indicating whether the matrix should be horizontal (TRUE) or vertical (FALSE)}
}
\value{
A matrix representing the combined array
}
\description{
This function generates a combined array matrix based on the specified parameters.
}
