% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{result_matrix}
\alias{result_matrix}
\title{Calculate cell counts within each spot based on combined array}
\usage{
result_matrix(combined_array)
}
\arguments{
\item{combined_array}{A matrix representing the spatial distribution of cells, with dimensions 40x20}
}
\value{
A matrix containing cell counts within each spot, with dimensions 40x20
}
\description{
This function takes a combined array representing the spatial distribution of cells
and calculates the cell counts within each spot based on randomly sampled values.
The input combined array should have dimensions of 40x20, representing a grid of 40x20 spots.
Each element of the combined array indicates the presence (1) or absence (0) of cells in the corresponding spot.
The function randomly samples a cell count between 5 and 20 for each spot and multiplies it with the combined array
to obtain the result matrix representing cell counts for each spot.
}
