% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{spatial_spot_nUMI}
\alias{spatial_spot_nUMI}
\title{Calculate total nUMI for all cell types}
\usage{
spatial_spot_nUMI(
  sim_spatial_spot_true_prop_data,
  result_matrices,
  cluster_names,
  org_data,
  prefix = ""
)
}
\arguments{
\item{result_matrices}{List of matrices representing the cell counts for each cell type}

\item{cluster_names}{List of names of the cell types}

\item{org_data}{Original data frame containing cluster information}

\item{spot_nUMI_filename}{Name of the CSV file to write the spatial spot nUMI counts}

\item{new_data_filename}{Name of the CSV file to write the new data}
}
\value{
A data frame containing the spatial spot nUMI counts for all specified cell types
}
\description{
This function calculates the total nUMI for all specified cell types and writes the result to CSV files.
}
